requirements.txt
# LiteNews AI Dependencies
# Install with: pip install -r requirements.txt

# Core LangGraph/LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langsmith>=0.1.0

# LLM Integrations
langchain-perplexity>=1.1.0
langchain-qwq>=0.1.0

# Search
langchain-tavily>=0.1.0
langchain-community>=0.3.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP
httpx>=0.27.0

# Development (optional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
